AllenNeuropixelsBase._getlfp — Method

Get the lfp data for a probe, providing indices for channels and times. See function below for indexing by channel ids and time values/intervals

AllenNeuropixelsBase.alignlfp — Method

For flashes alignment, trail=false will return only the data from within the flash period. trail=onset will return the data from the onset of the flash to the onset of the flash through to the onset of the next flash. trail=offset will return the data from the offset of the flash to the onset of the next flash.

AllenNeuropixelsBase.formatspikes — Method

A function to easily grab formatted spike data for a given session, using some sensible default parameters

AllenNeuropixelsBase.getdownsampledlfp — Method

If you want to downsample the LFP data, its quicker to use this function and then perform slicing afterwards (since getlfp() has to check all of the time coordinates you supply, which can be slow).

AllenNeuropixelsBase.getlfp — Method

Now we can overload getlfp() to index by structure

AllenNeuropixelsBase.getlfp — Method

This is the one you should be using. Get lfp data by channel id and time intervals or vector. Also, throw error if you try to access an invalid time interval.

AllenNeuropixelsBase.getprobes — Method

`getprobes()`

Read the probe data from the EcephysProjectCache object returned by ecephyscache() and return it as a DataFrame.

Returns

A DataFrame containing the probe data.

AllenNeuropixelsBase.getsessiontable — Method

getsessiontable — Read and return the session table from the EcephysProjectCache

`getsessiontable()`

Read the session table data from the EcephysProjectCache object returned by ecephyscache and returns it as a DataFrame.

AllenNeuropixelsBase.getspikes — Function

We combine the spike times and spike amplitudes into one sparse array, using a given bin width.

AllenNeuropixelsBase.interpmatch — Method

Match the time indices of the first input DimVector to the second by interpolating the first

AllenNeuropixelsBase.loaddataframe — Function

loaddataframe(file, dir=datadir)::DataFrame

Load a CSV file into a DataFrame.

Arguments:

file: A string representing the name of the CSV file to be loaded.

dir (optional): A string representing the directory containing the CSV file. Default is datadir.

Returns:

A DataFrame object containing the contents of the CSV file.

Example:

using DataFrames
df = loaddataframe("mydata.csv", "/path/to/my/data/")

AllenNeuropixelsBase.matchlfp — Method

Adjust the times of LFP matrix Y so that they match the matrix X

AllenNeuropixelsBase.spikematrix — Function

Construct a sparse array of spike counts from a Dict of spike times

AllenNeuropixelsBase.VisualBehavior.datapaths — Method

 datapaths()

Get the paths to the visual behavior neuropixels data on Amazon S3 as a tree-like dictionary.

AllenNeuropixelsBase.VisualBehavior.filetree — Method

filetree(paths::Vector{String})

Creates a dictionary-like object representing the directory and file structure from a list of file paths.

Arguments:

paths: A vector of file paths to include in the tree.

Returns:

dict_tree : A dictionary-like object representing the directory and file structure.

AllenNeuropixelsBase.VisualBehavior.getmanifesturl — Function

getmanifesturl(manifest_version="0.5.0")

Return the URL for the neuropixels visual behavior project manifest file hosted on a given hostname server and version manifest_version.

Arguments

manifest_version::String: The version of the manifest file. Default value is "0.5.0".

Other manifest version numbers can be identified here: https://s3.console.aws.amazon.com/s3/buckets/visual-behavior-neuropixels-data?prefix=visual-behavior-neuropixels%2Fmanifests%2F&region=us-west-2

AllenNeuropixelsBase.VisualBehavior.getsessionfiles — Method

getsessionfiles()

Return a dictionary mapping session IDs to their corresponding file paths.